test(feeds): align sorting helpers with Long sortOrder
- Use `0L` as the fallback sort key in feed sorting.
- Update test helper `sortOrder` type to `Long?`.

+++ app/src/test/java/me/ash/reader/domain/data/FeedSortingTest.kt
@@ -148,7 +148,7 @@ class FeedSortingTest {
     private fun testFeed(
         id: String,
         name: String = "Test Feed",
-        sortOrder: Int? = null,
+        sortOrder: Long? = null,
         groupId: String = "1\$Defaults",
         accountId: Int = 1,
     ) = Feed(
